Refactored LandingPage file structure. Will likely change based on new UX spec
