+++ src/components/pages/LandingPage/SignupForm/index.ts
@@ -1,0 +1,1 @@
+export { default as Signup } from './SignupFormContainer';
