+++ src/components/pages/LandingPage/Splash/LoginSpash.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const LoginSplash = (): React.ReactElement => {
+  return (
+    <h2>
+      Sign in for the <span className="b">Free Daily Story Contest!</span>
+    </h2>
+  );
+};
+
+export default LoginSplash;
