+++ src/components/pages/LandingPage/Splash/index.ts
@@ -1,0 +1,1 @@
+export { default as Splash } from './LandingSplash';
